\docType{package}
\name{RMRAINGEN-package}
\alias{RMRAINGEN}
\alias{RMRAINGEN-package}
\title{R - Multi-site RAIN Generator}
\description{
RMRAINGEN: R Multi-site RAIN GENerator
}
\details{
This package contains functions and S3 methods for spatial
multi-site stochastic generation of daily precipitation.
Bugs/comments/questions/collaboration of any kind are
warmly welcomed.

\tabular{ll}{ Package: \tab RMARAINGEN\cr Type: \tab
Package\cr Version: \tab 1.0\cr Date: \tab 2014-07-22\cr
License: \tab GPL (>= 2)

\cr LazyLoad: \tab yes\cr Depends: see DESCRIPTION file \cr
}
}
\note{
RMRAINGEN is free software: you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation, either
version 3 of the License, or (at your option) any later
version.

RMRAINGEN is distributed in the hope that it will be
useful, but WITHOUT ANY WARRANTY; without even the implied
warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
PURPOSE.  See the GNU General Public License for more
details.

You should have received a copy of the GNU General Public
License along with this program.  If not, see
<http://www.gnu.org/licenses/>.
}
\author{
Emanuele Cordano \email{emanuele.cordano@gmail.org},
Emanuele Eccel \email{emanuele.eccel@iasma.it}
}
\keyword{package}

